Notes Clearing and Cache issues resolved
Http Decorator now clears a cache item if the cache config is
explictly set to false. This allows a new cache item to come in and
update the value with something more recent.

Lastly we duplicate the notes data to keep the markers and listing
separate to not see the visual clear that ui-leaflet requires.